update Ibase Services
Change IBaseService methods to use string IDs

Updated the IBaseService interface in the Turo.Application.Interfaces namespace to change the parameter type of the GetByIdAsync and DeleteAsync methods from int to string. This change reflects a shift from numeric to string-based identifiers for better flexibility and alignment with application requirements.