logiuqe base service update
Change DeleteAsync parameter type from int to string

The `DeleteAsync` method in the `BaseService` class was modified to change the type of its parameter from `int` to `string`. This change affects how the method retrieves and deletes entities based on their identifier, which is now expected to be a string rather than an integer.